OGC Refactoring Part 12 - Rename ogc packages to lower case - Rename package ogcApi to ogc_api_records - Rename package researchQuery to research_query - Rename package exportCatalog to export_catalog - Rename package exportRecord to export_record
