+++ server/src/main/java/de/ingrid/igeserver/features/ogc_api_records/export_record/OgcGeoJsonExporter.kt
@@ -17,41 +17,40 @@
  * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
-package de.ingrid.igeserver.features.ogcApi.exportRecord
+package de.ingrid.igeserver.features.ogc_api_records.export_record
 
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.node.ObjectNode
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
+import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import de.ingrid.igeserver.exports.ExportOptions
 import de.ingrid.igeserver.exports.ExportTypeInfo
 import de.ingrid.igeserver.exports.IgeExporter
-import de.ingrid.igeserver.features.ogcApi.services.OgcHtmlConverterService
+import de.ingrid.igeserver.features.ogc_api_records.model.Record
+import de.ingrid.igeserver.features.ogc_api_records.model.RecordTime
 import de.ingrid.igeserver.persistence.postgresql.jpa.model.ige.Document
-import de.ingrid.igeserver.services.CatalogService
-import de.ingrid.igeserver.services.DocumentCategory
-import de.ingrid.igeserver.services.DocumentService
+import de.ingrid.igeserver.services.*
 import org.springframework.context.annotation.Lazy
 import org.springframework.http.MediaType
 import org.springframework.stereotype.Service
 
-
 @Service
-class OgcHtmlExporter(
+class OgcGeoJsonExporter(
         @Lazy val documentService: DocumentService,
         val catalogService: CatalogService,
-        val ogcHtmlConverterService: OgcHtmlConverterService
 ) : IgeExporter {
 
-    override val typeInfo = ExportTypeInfo(
-            DocumentCategory.DATA,
-            "html",
-            "htmlIGE",
-            "HTML Export des IGE",
-            MediaType.TEXT_HTML_VALUE,
-            "text/html",
-            emptyList(),
-            false
-    )
-
+    override val typeInfo: ExportTypeInfo
+        get() = ExportTypeInfo(
+                DocumentCategory.DATA,
+                "geojson",
+                "geoJsonIGE",
+                "geoJson Export des IGE",
+                MediaType.APPLICATION_JSON_VALUE,
+                "json",
+                listOf(),
+                false
+        )
     override fun run(doc: Document, catalogId: String, options: ExportOptions): Any {
         // TODO: move to utilities to prevent cycle
         val version = documentService.convertToJsonNode(doc)
@@ -63,13 +62,11 @@ class OgcHtmlExporter(
             Pair(version, null)
         }
 
-        val exportData = selectDraftOrPublished(versions.first, versions.second)
-        return ogcHtmlConverterService.convertObjectNode2Html(exportData, null)
-    }
-
-    fun runCollection(catalogId: String) {
-        val catalog = catalogService.getCatalogById(catalogId)
-        return
+        val dataset = versions.first ?: versions.second
+        val isAddress: Boolean = if(dataset?.get("address") == null) false else true
+        if(isAddress) return dataset as Any
+        val record = addExportWrapper(catalogId, versions.first, versions.second)
+        return convertToJsonNode(record)
     }
 
     private fun getPublished(catalogId: String, uuid: String): JsonNode? {
@@ -82,13 +79,44 @@ class OgcHtmlExporter(
         }
     }
 
-    private fun selectDraftOrPublished(publishedVersion: JsonNode?, draftVersion: JsonNode?): ObjectNode {
-        return (draftVersion ?: publishedVersion) as ObjectNode
+    private fun addExportWrapper(catalogId: String, publishedVersion: JsonNode?, draftVersion: JsonNode?): Record {
+
+        val dataset = publishedVersion ?: draftVersion
+
+        val geometry: MutableList<JsonNode> = mutableListOf()
+        val interval: MutableList<String> = mutableListOf()
+        interval.add("..") // start date
+        interval.add("..") // end date
+        val time = RecordTime( interval, resolution = "P1D" )
+
+        return Record(
+                id = dataset!!.get("_uuid"),
+                conformsTo = null,
+                type = "Feature",
+                time,
+                geometry,
+                properties = dataset
+        )
     }
 
+    fun convertToJsonNode(record: Record): JsonNode {
+
+        val mapper = jacksonObjectMapper()
+        mapper.registerModule(JavaTimeModule())
+
+        val node = mapper.convertValue(record, ObjectNode::class.java)
+        val data = node
+        data.fields().forEach { entry ->
+            node.replace(entry.key, entry.value)
+        }
+        return node
+
+    }
 
     override fun toString(exportedObject: Any): String {
         return exportedObject as String
     }
+
+
 
 }
