+++ server/src/main/java/de/ingrid/igeserver/features/ogc_api_records/model/OgcRecord.kt
@@ -17,7 +17,7 @@
  * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
-package de.ingrid.igeserver.features.ogcApi.model
+package de.ingrid.igeserver.features.ogc_api_records.model
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.databind.JsonNode
