+++ server/src/main/java/de/ingrid/igeserver/features/ogc_api_records/services/research_query/OgcApiResearchQueryFactory.kt
@@ -17,7 +17,7 @@
  * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
-package de.ingrid.igeserver.features.ogcApi.services.researchQuery
+package de.ingrid.igeserver.features.ogc_api_records.services.research_query
 
 import de.ingrid.igeserver.configuration.ConfigurationException
 import de.ingrid.igeserver.model.ResearchQuery
